itsm-api/requirements.txt
fastapi>=0.115.0
celery[redis]>=5.4.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
httpx>=0.27.2
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
python-multipart>=0.0.19
redis>=5.2.0
slowapi>=0.1.9
limits>=5.0
prometheus-fastapi-instrumentator>=7.0
python-magic>=0.4.27
jinja2>=3.1.0
Pillow>=10.0.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
openpyxl>=3.1.0
minio>=7.2.0
